Moved file, added to src.index.js
