+++ src/index.js
@@ -27,3 +27,4 @@ export AccordionToggle from './Collapsible/Accordion/AccordionToggle'
 export Collapsible from './Collapsible/Collapsible/Collapsible'
 export CollapsibleBody from './Collapsible/Collapsible/CollapsibleBody'
 export CollapsibleToggle from './Collapsible/Collapsible/CollapsibleToggle'
+export Border from './Border'
